para persistence: AlumnoDao, AsignaturaDao, MateriaDao y ProfesorDao agregadas

+++ src/main/java/ar/edu/utn/frbb/tup/business/AlumnoService.java
@@ -15,7 +15,7 @@ public interface AlumnoService {
 
     Alumno modificarAlumno(long idAlumno, AlumnoDto alumno);
 
-    Alumno buscarAlumno(int id);
+    Alumno buscarAlumno(long id);
 
     // ------------------------------------------------------------------------
